[ChessRouler] More Refinements
