+++ util-kit/src/lib/ChessFENBoard/util/board.ts
@@ -141,63 +141,3 @@ export const chessBoardToFenBoard = (chessBoard: ChessBoard): FENBoard =>
 
     return (p.color === 'b' ? p.type : p.type.toUpperCase()) as PieceSymbol;
   });
-
-// export const pgnToFen = (pgn: ChessPGN) => {
-//   const instance = new Chess();
-
-//   // instance.
-
-//   instance.loadPgn(pgn);
-
-//   return instance.fen();
-// };
-
-// export const BITS: Record<string, number> = {
-//   NORMAL: 1,
-//   CAPTURE: 2,
-//   BIG_PAWN: 4,
-//   EP_CAPTURE: 8,
-//   PROMOTION: 16,
-//   KSIDE_CASTLE: 32,
-//   QSIDE_CASTLE: 64,
-// }
-
-// export const chessNoveToSan = (move: Move): ChessMoveSan => {
-//   let output = '';
-
-//   if ((move.flags as any) & BITS['KSIDE_CASTLE']) {
-//     output = 'O-O'
-//   } else if ((move.flags as any) & BITS['QSIDE_CASTLE']) {
-//     output = 'O-O-O'
-//   } else {
-//     if (move.piece !== 'p') {
-//       const disambiguator = getDisambiguator(move, moves);
-//       output += move.piece.toUpperCase() + disambiguator
-//     }
-
-//     if ((move.flags as any) & (BITS['CAPTURE'] | BITS['EP_CAPTURE'])) {
-//       if (move.piece === 'p') {
-//         output += algebraic(move.from)[0]
-//       }
-//       output += 'x'
-//     }
-
-//     output += algebraic(move.to)
-
-//     if (move.promotion) {
-//       output += '=' + move.promotion.toUpperCase()
-//     }
-//   }
-
-//   this._makeMove(move)
-//   if (this.isCheck()) {
-//     if (this.isCheckmate()) {
-//       output += '#'
-//     } else {
-//       output += '+'
-//     }
-//   }
-//   this._undoMove()
-
-//   return output
-// }
