+++ util-kit/src/lib/ChessRouler/lib/chessColor/index.ts
@@ -1,0 +1,2 @@
+export * from './chessColor';
+export * from './types';
