+++ util-kit/src/lib/ChessRouler/types.ts
@@ -1,5 +1,6 @@
 import type { Square, Move, PieceSymbol } from 'chess.js';
 import { FenBoardPromotionalPieceSymbol } from '../ChessFENBoard';
+import { ChessColor } from './lib/chessColor/types';
 
 // Remove in favor of ShortChessMove
 export type ChessMove = {
@@ -33,30 +34,11 @@ export type ChessFENStateNotation = ChessFENStateNotationBrand & string;
 
 export type ChessMoveSan = string; // TODO: Brand this type
 
-export type WhiteShortColor = 'w';
-export type BlackShortColor = 'b';
-
-// @deprecate
-export type WhiteLongColor = 'white';
-export type BlackLongColor = 'black';
-
-export type WhiteColor = WhiteShortColor;
-// export type WhiteColor = WhiteShortColor | WhiteLongColor;
-export type BlackColor = BlackShortColor;
-// export type BlackColor = BlackShortColor | BlackLongColor;
-
-// @deprecate
-// export type LongChessColor = WhiteShortColor | BlackShortColor;
-export type LongChessColor = WhiteLongColor | BlackLongColor;
-export type ShortChessColor = WhiteShortColor | BlackShortColor;
-
 export type ChessSide = 'home' | 'away';
-
-export type ChessColor = WhiteColor | BlackColor;
 
 export type DetailedChessPiece = {
   piece: PieceSymbol;
-  color: ShortChessColor;
+  color: ChessColor;
 };
 
 export type PieceSan =
@@ -89,5 +71,5 @@ export enum GameOverReason {
   'resignation',
   'acceptedDraw',
   'aborted',
-  'drawInsufficientMaterial'
+  'drawInsufficientMaterial',
 }
